+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add text stroke class to Text component
 - Add close-circle and arrow-down-arrow-up icons
 - Add `reload` option to Countdown component
+- Add `offset-end` option to Sticky component
 
 ### Changed